feat: documentação da api em swagger

+++ app/Http/Controllers/Api/UserController.php
@@ -16,6 +16,22 @@ class UserController extends Controller
         return $this->user = $user;
     }
 
+    /**
+    *  @OA\GET(
+    *      path="/api/user",
+    *      summary="Get all users",
+    *      description="Get all users",
+    *      tags={"Users"},
+    *      @OA\Response(
+    *          response=200,
+    *          description="OK",
+    *          @OA\MediaType(
+    *              mediaType="application/json",
+    *          )
+    *      ),
+    *
+    *  )
+    */
     public function index() {
         return response()->json($this->user->select(['id', 'name', 'email', 'company', 'created_at', 'updated_at'])->with('phones:user_id,num')->get(), 200);
     }
@@ -25,7 +41,29 @@ class UserController extends Controller
      *
      * @param  Integer
      * @return \Illuminate\Http\Response
-     */
+     *
+     *  @OA\GET(
+     *      path="/api/user/{id}",
+     *      summary="Get user",
+     *      description="Get user",
+     *      tags={"Users"},
+     *      @OA\Parameter(
+     *         name="id",
+     *         in="path",
+     *         description="ID of the user",
+     *         required=true,
+     *         @OA\Schema(type="integer")
+     *     ),
+     *      @OA\Response(
+     *          response=200,
+     *          description="OK",
+     *          @OA\MediaType(
+     *              mediaType="application/json",
+     *          )
+     *      ),
+     *
+     *  )
+    */
     public function show($id) {
         $user = $this->user->select(['id', 'name', 'image', 'email', 'company', 'created_at', 'updated_at'])->with('phones:user_id,num')->find($id);
 
@@ -36,6 +74,68 @@ class UserController extends Controller
         return response()->json($user, 200);
     }
 
+    /**
+    *  @OA\POST(
+    *      path="/api/user",
+    *      summary="Create a new user",
+    *      description="Create a new user by providing the necessary information",
+    *      tags={"Users"},
+    *      @OA\RequestBody(
+    *         required=true,
+    *         @OA\MediaType(
+    *             mediaType="multipart/form-data",
+    *             @OA\Schema(
+    *                 required={"name", "email", "company","password", "image"},
+    *                 @OA\Property(
+    *                     property="name",
+    *                     type="string",
+    *                     description="User's name"
+    *                 ),
+    *                 @OA\Property(
+    *                     property="email",
+    *                     type="string",
+    *                     format="email",
+    *                     description="User's email"
+    *                 ),
+    *                 @OA\Property(
+    *                     property="company",
+    *                     type="string",
+    *                     description="User's company"
+    *                 ),
+    *                 @OA\Property(
+    *                     property="password",
+    *                     type="string",
+    *                     format="password",
+    *                     description="User's password"
+    *                 ),
+    *                 @OA\Property(
+    *                     property="image",
+    *                     type="string",
+    *                     format="binary",
+    *                     description="User profile image (PNG, JPEG, JPG)"
+    *                 )
+    *             )
+    *         )
+    *     ),
+    *      @OA\Response(
+    *          response=201,
+    *          description="User created successfully",
+    *          @OA\MediaType(
+    *              mediaType="application/json",
+    *              @OA\Schema(
+    *                  type="object",
+    *                  @OA\Property(property="id", type="integer", example=1),
+    *                  @OA\Property(property="image", type="string", example="images/jsko99asd00dj.jpg"),
+    *                  @OA\Property(property="name", type="string", example="John Doe"),
+    *                  @OA\Property(property="email", type="string", example="john.doe@example.com"),
+    *                  @OA\Property(property="company", type="string", example="Microsoft"),
+    *                  @OA\Property(property="created_at", type="string", format="date-time", example="2024-12-07T12:00:00Z"),
+    *                  @OA\Property(property="updated_at", type="string", format="date-time", example="2024-12-07T12:00:00Z")
+    *              )
+    *          )
+    *      ),
+    *  )
+    */
     public function store(Request $request) {
 
         $request->validate($this->user->rules());
@@ -56,7 +156,7 @@ class UserController extends Controller
             return response()->json($e, 500);
         }
 
-        return response()->json($user, 201);
+        return response()->json($user->makeHidden('password'), 201);
     }
 
     /**
@@ -66,6 +166,44 @@ class UserController extends Controller
      * @param  Integer 
      * @return \Illuminate\Http\Response
      */
+    /**
+    *  @OA\PATCH(
+    *      path="/api/user/{id}",
+    *      summary="Update a new user",
+    *      description="Update a new user by providing the necessary information",
+    *      tags={"Users"},
+    *      @OA\Parameter(
+    *          name="id",
+    *          in="path",
+    *          required=true,
+    *          description="user id",
+    *          @OA\Schema(type="integer")
+    *      ),
+    *      @OA\RequestBody(
+    *          @OA\JsonContent(
+    *              type="object",
+    *              @OA\Property(property="name", type="string", example="John Doe", description="The user's name"),
+    *              @OA\Property(property="email", type="string", format="email", example="john.doe@example.com", description="The user's email"),
+    *              @OA\Property(property="company", type="string", example="Microsoft", description="Company you work for"),
+    *          )
+    *      ),
+    *      @OA\Response(
+    *          response=200,
+    *          description="User updated successfully",
+    *          @OA\MediaType(
+    *              mediaType="application/json",
+    *              @OA\Schema(
+    *                  type="object",
+    *                  @OA\Property(property="id", type="integer", example=1),
+    *                  @OA\Property(property="name", type="string", example="John Doe"),
+    *                  @OA\Property(property="email", type="string", example="john.doe@example.com"),
+    *                  @OA\Property(property="company", type="string", example="Microsoft"),
+    *                  @OA\Property(property="created_at", type="string", format="date-time", example="2024-12-07T12:00:00Z")
+    *              )
+    *          )
+    *      ),
+    *  )
+    */
     public function update(Request $request, $id)
     {
         $user = $this->user->find($id);
@@ -99,7 +237,7 @@ class UserController extends Controller
 
         $user->save();
 
-        return response()->json($user, 200);
+        return response()->json($user->makeHidden('password'), 200);
     }
 
     /**
@@ -108,6 +246,28 @@ class UserController extends Controller
      * @param  Integer
      * @return \Illuminate\Http\Response
      */
+    /**
+     *  @OA\DELETE(
+     *      path="/api/user/{id}",
+     *      summary="Delete user",
+     *      description="Delete user",
+     *      tags={"Users"},
+     *      @OA\Parameter(
+     *         name="id",
+     *         in="path",
+     *         description="user id",
+     *         required=true,
+     *      ),
+     *      @OA\Response(
+     *          response=200,
+     *          description="User removed successfully.",
+     *          @OA\MediaType(
+     *              mediaType="application/json",
+     *          )
+     *      ),
+     *
+     *  )
+    */
     public function destroy($id)
     {
         $user = $this->user->find($id);
@@ -116,7 +276,9 @@ class UserController extends Controller
             return response()->json(['error' => 'Resource searched for does not exist.'], 404);
         }
 
-        Storage::disk('public')->delete($user->image);
+        if ($user->image !== null) {
+            Storage::disk('public')->delete($user->image);
+        }
 
         $user->phones()->delete();
         $user->delete();
